requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.25.0
GitPython>=3.1.0
jinja2>=3.0.0
